fix: async logic in producer

+++ simulated-gps-producer/src/schema.py
@@ -40,7 +40,7 @@ simulated_gps_data_schema: T.StructType = T.StructType([
     T.StructField("hour", T.IntegerType()),
     T.StructField("minute", T.IntegerType()),
     T.StructField("second", T.IntegerType()),
-    T.StructField("microsecond", T.IntegerType()),
+    T.StructField("microsecond", T.LongType()),
     T.StructField("fare_amount", T.DoubleType()),
     T.StructField("tip_amount", T.DoubleType()),
     T.StructField("total_profit", T.DoubleType()),
